[chore][graph] Separate node types
