+++ service/internal/graph/consumer.go
@@ -1,0 +1,17 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package graph // import "go.opentelemetry.io/collector/service/internal/graph"
+
+import (
+	"go.opentelemetry.io/collector/consumer"
+)
+
+// baseConsumer redeclared here since not public in consumer package. May consider to make that public.
+type baseConsumer interface {
+	Capabilities() consumer.Capabilities
+}
+
+type consumerNode interface {
+	getConsumer() baseConsumer
+}
