+++ service/internal/graph/node.go
@@ -1,0 +1,21 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package graph // import "go.opentelemetry.io/collector/service/internal/graph"
+
+import (
+	"hash/fnv"
+	"strings"
+)
+
+type nodeID int64
+
+func (n nodeID) ID() int64 {
+	return int64(n)
+}
+
+func newNodeID(parts ...string) nodeID {
+	h := fnv.New64a()
+	h.Write([]byte(strings.Join(parts, "|")))
+	return nodeID(h.Sum64())
+}
